requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
Django==5.2.8
django-browser-reload==1.21.0
django-tailwind==4.4.2
djangorestframework==3.16.1
exif==1.6.1
honcho==2.0.0
idna==3.11
Jinja2==3.1.6
kml2geojson==5.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.0.0
pillow_heif==1.1.1
plum-py==0.8.7
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.4
text-unidecode==1.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.6.2
whitenoise==6.11.0
